Refactor(docs): Refactor storybook utils
